refactor: split _op_function_bodies.py into per-function files
One file per function matching causal_conv.py and linear_attention.py pattern:
- skip_layer_normalization.py: SkipLayerNormalization + SkipSimplifiedLayerNormalization
- simplified_layer_normalization.py: SimplifiedLayerNormalization

Registry helpers (get_function, register_function_bodies, _FUNCTION_BUILDERS)
moved to functions/__init__.py where all exports are aggregated.

_optimizations.py updated to import register_function_bodies from
mobius.functions (not the deleted _op_function_bodies module).

2275 tests pass.
